+++ lua/rustaceanvim/runnables.lua
@@ -116,6 +116,29 @@ function M.run_command(choice, runnables)
   end
 end
 
+---@param runnable RARunnable
+---@return boolean
+local function is_testable(runnable)
+  ---@cast runnable RARunnable
+  local cargoArgs = runnable.args and runnable.args.cargoArgs or {}
+  return #cargoArgs > 0 and vim.startswith(cargoArgs[1], 'test')
+end
+
+---@param executableArgsOverride? string[]
+---@param runnables RARunnable[]
+---@return RARunnable[]
+function M.apply_exec_args_override(executableArgsOverride, runnables)
+  if type(executableArgsOverride) == 'table' and #executableArgsOverride > 0 then
+    local unique_runnables = {}
+    for _, runnable in pairs(runnables) do
+      runnable.args.executableArgs = executableArgsOverride
+      unique_runnables[vim.inspect(runnable)] = runnable
+    end
+    runnables = vim.tbl_values(unique_runnables)
+  end
+  return runnables
+end
+
 ---@param executableArgsOverride? string[]
 ---@param opts RunnablesOpts
 ---@return fun(_, result: RARunnable[])
@@ -125,20 +148,9 @@ local function mk_handler(executableArgsOverride, opts)
     if runnables == nil then
       return
     end
-    if type(executableArgsOverride) == 'table' and #executableArgsOverride > 0 then
-      local unique_runnables = {}
-      for _, runnable in pairs(runnables) do
-        runnable.args.executableArgs = executableArgsOverride
-        unique_runnables[vim.inspect(runnable)] = runnable
-      end
-      runnables = vim.tbl_values(unique_runnables)
-    end
+    runnables = M.apply_exec_args_override(executableArgsOverride, runnables)
     if opts.tests_only then
-      runnables = vim.tbl_filter(function(runnable)
-        ---@cast runnable RARunnable
-        local cargoArgs = runnable.args and runnable.args.cargoArgs or {}
-        return #cargoArgs > 0 and vim.startswith(cargoArgs[1], 'test')
-      end, runnables)
+      runnables = vim.tbl_filter(is_testable, runnables)
     end
 
     -- get the choice from the user
@@ -157,6 +169,56 @@ local function mk_handler(executableArgsOverride, opts)
   end
 end
 
+---@param position lsp.Position
+---@param targetRange lsp.Range
+local function is_within_range(position, targetRange)
+  return targetRange.start.line <= position.line and targetRange['end'].line >= position.line
+end
+
+---@param runnables RARunnable
+---@return integer | nil choice
+function M.get_runnable_at_cursor_position(runnables)
+  ---@type lsp.Position
+  local position = vim.lsp.util.make_position_params().position
+  ---@type integer|nil, integer|nil
+  local choice, fallback
+  for idx, runnable in ipairs(runnables) do
+    if runnable.location then
+      local range = runnable.location.targetRange
+      if is_within_range(position, range) then
+        if vim.startswith(runnable.label, 'test-mod') then
+          fallback = idx
+        else
+          choice = idx
+          break
+        end
+      end
+    end
+  end
+  return choice or fallback
+end
+
+local function mk_cursor_position_handler(executableArgsOverride)
+  ---@param runnables RARunnable[]
+  return function(_, runnables)
+    if runnables == nil then
+      return
+    end
+    runnables = M.apply_exec_args_override(executableArgsOverride, runnables)
+    local choice = M.get_runnable_at_cursor_position(runnables)
+    if not choice then
+      vim.notify('No runnable targets found for the current position.', vim.log.levels.ERROR)
+      return
+    end
+    M.run_command(choice, runnables)
+    local cached_commands = require('rustaceanvim.cached_commands')
+    if is_testable(runnables[choice]) then
+      cached_commands.set_last_testable(choice, runnables)
+    end
+    cached_commands.set_last_runnable(choice, runnables)
+  end
+end
+
 ---@class RunnablesOpts
 ---@field tests_only? boolean
 
@@ -167,6 +229,10 @@ function M.runnables(executableArgsOverride, opts)
   ---@type RunnablesOpts
   opts = vim.tbl_deep_extend('force', { tests_only = false }, opts or {})
   vim.lsp.buf_request(0, 'experimental/runnables', get_params(), mk_handler(executableArgsOverride, opts))
+end
+
+function M.run(executableArgsOverride)
+  vim.lsp.buf_request(0, 'experimental/runnables', get_params(), mk_cursor_position_handler(executableArgsOverride))
 end
 
 return M